Fine tune segmentation of cell body
Key element now is to process each cell individually and fill in with closing until there is a change of the euler numbers. This works well  when there are some cells that are a bit open (like a letter C) and then it closes, creates a hole, that eventually becomes closed.